Adresse les retours de la PR review
- Extrait PersonIcon vers components/icons/person.tsx (évite la duplication avec GenderDistributionChart)
- Split server/client : page.tsx reste Server Component, état des onglets isolé dans AutresTabs.tsx
- ARIA tabs complets : role=tablist, role=tab, aria-selected, aria-controls, role=tabpanel, hidden
- Dates lues depuis pouvoir.json (plus de literals hardcodés dans le JSX)
- role=img + aria-label sur le conteneur des icônes personnes
- TODO ajouté sur les données provisoires et les logos manquants
